% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silva128.1MgDb.R
\docType{package}
\name{silva128.1MgDb-package}
\alias{silva128.1MgDb-package}
\title{SILVA SSU rRNA database MgDb Package.}
\description{
This package loads a MgDb class object with
sequence and taxonomic data for the SILVA
SSU rRNA database release 128.1, \url{https://www.arb-silva.de}
}
\examples{
ls('package:silva128.1MgDb')
slv128.1MgDb
}
\seealso{
\link[metagenomeFeatures]{MgDb-class}
}
\author{
Nathan D. Olson \email{nolson@umiacs.umd.edu}
}
\keyword{package}
