% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slv128.1MgDb.R
\name{slv128.1MgDb}
\alias{slv128.1MgDb}
\title{MgDb-class object for SILVA SSU rRNA release 128.1}
\description{
slv128.1MgDb is a MgDb-class object with
sequence and taxonomic data for the SILVA SSU rRNA
database version 128.1, \url{https://www.arb-silva.de/}
}
\examples{
slv128.1MgDb
}
\seealso{
\link[metagenomeFeatures]{MgDb-class}
}
